+++ test/snippet/utility/views/convert_int_to_bool.cpp
@@ -1,10 +1,10 @@
+#include <seqan3/std/ranges>
 #include <vector>
 
 #include <seqan3/alphabet/nucleotide/dna15.hpp>
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/range/views/convert.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/views/convert.hpp>
 
 int main()
 {